Move code to scr folder
All code is moved to the scr folder to clean up the root dir somewhat.
